+++ services/intersection-api/api/src/main/java/us/dot/its/jpo/ode/api/models/postgres/tables/MaxRetryLimitReachedInstance.java
@@ -1,26 +1,26 @@
 package us.dot.its.jpo.ode.api.models.postgres.tables;
 
-import jakarta.persistence.EmbeddedId;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.MapsId;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
 @Entity
-@Table(name = "max_retry_limit_reached_instances", schema = "public")
+@Table(name = "max_retry_limit_reached_instances")
 public class MaxRetryLimitReachedInstance {
-  @EmbeddedId
-  private MaxRetryLimitReachedInstanceId id;
+    @EmbeddedId
+    private Integer id;
 
-  @MapsId
-  @ManyToOne(fetch = FetchType.LAZY, optional = false)
-  @JoinColumn(name = "rsu_id", nullable = false)
-  private Rsu rsu;
+    @MapsId("rsuId")
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "rsu_id", nullable = false)
+    private Rsu rsu;
+
+    @NotNull
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "target_firmware_version", nullable = false)
+    private FirmwareImage targetFirmwareVersion;
 
 }
