+++ services/intersection-api/api/src/main/java/us/dot/its/jpo/ode/api/models/postgres/tables/SpatialRefSy.java
@@ -1,4 +1,4 @@
-package us.dot.its.jpo.ode.api;
+package us.dot.its.jpo.ode.api.models.postgres.tables;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
